rename pkg/server to pkg/grpc
