+++ pkg/grpc/options.go
@@ -1,4 +1,4 @@
-package server
+package grpc
 
 import (
 	"fmt"
